feat: impede remoção de tipo de solo em uso

+++ src/soil-types/soil-types.service.ts
@@ -1,12 +1,16 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { SoilTypesDto } from './dto/soil-types.dto';
 import { SoilTypesRepository } from './soil-types.repository';
 import { SoilTypes } from './entities/soil-types.entity';
 import { UpdateSoilTypeDto } from './update-soil-type.dto';
+import { AreasRepository } from 'src/areas/areas.repository';
 
 @Injectable()
 export class SoilTypesService {
-  constructor(private readonly repository: SoilTypesRepository) { }
+  constructor(
+    private readonly repository: SoilTypesRepository,
+    private readonly areasRepository: AreasRepository,
+  ) { }
 
   async create(soilTypesDto: SoilTypesDto): Promise<SoilTypes> {
     const data = {
@@ -39,6 +43,12 @@ export class SoilTypesService {
     if (!soilType) {
       throw new NotFoundException(`Tipo de solo com o ID #${id} não encontrado.`);
     }
+
+    const areaInUse = await this.areasRepository.existsBySoilTypeId(id);
+    if (areaInUse) {
+      throw new BadRequestException('Não é possível remover! tipo de solo está em uso por uma ou mais áreas.');
+    }
+
     return this.repository.remove(id);
   }
 }
